requirements.txt
pandas
snowflake-connector-python
dbt-snowflake
dbt-core
